pyproject.toml
[project]
name = "itmo-mlops"
version = "0.1.0"
description = ""
authors = [
    {name = "Georgy Kiselev", email = "gsx2002@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.11.0,<3.13"

dependencies = [
    "requests>=2.32.3,<3.0.0",
    "matplotlib>=3.1.0",
    "numpy>=2.2.5,<3.0.0",
    "ipykernel (>=6.29.5,<7.0.0)",
    "natsort (>=8.4.0,<9.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "huggingface-hub (>=0.31.1,<0.32.0)",
    "dvc (>=3.59.2,<4.0.0)",
    "dvc-gdrive (>=3.0.1,<4.0.0)",
    "rclone (>=0.4.4,<0.5.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "torchvision (>=0.20.0)",
    "opencv-python-headless (>=4.11.0.86,<5.0.0.0)",
    "scipy (>=1.15.3,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "scikit-image (>=0.25.2,<0.26.0)",
    "python-bidi (>=0.6.6,<0.7.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "shapely (>=2.1.0,<3.0.0)",
    "pyclipper (>=1.3.0.post6,<2.0.0)",
    "ninja (>=1.11.1.4,<2.0.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "torch (>=2.7.0)",
    "python-doctr[torch] (>=0.11.0,<0.12.0)",
    "snakemake (>=9.5.1,<10.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "validators (>=0.35.0,<0.36.0)",
    "clearml (>=2.0.0,<3.0.0)",
    "pyopenssl (==23.2.0)",
    "cryptography (==41.0.7)",
]

[tool.poetry.dependencies]
python = ">=3.11.0, <3.13"
torch = { version = "2.7.0+cu128", source = "torch-cu128" }

[[tool.poetry.source]]
name = "torch-cu128"
url = "https://download.pytorch.org/whl/cu128"
priority = "explicit"


[tool.poetry.group.app.dependencies]
wandb = ">=0.10.31"
fastapi = "^0.104.1"
pydantic = "^2.11.3"
pymongo = "^4.6.0"
python-magic = "^0.4.27"
wget = "^3.2"
uvicorn = ">=0.24.0"
python-multipart = "^0.0.6"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.0"
dvc = "^3.30.1"
python-doctr = { version = ">=0.11.0,<0.12.0", extras = ["torch"] }
pandas = "^2.1.3"
tqdm = "^4.66.1"
requests = "^2.31.0"
pyyaml = "^6.0.1"
pyOpenSSL = "<24.0.0"


[tool.poetry.group.ocr.dependencies]
requests = ">=2.32.3,<3.0.0"
random-word = ">=1.0.13,<2.0.0"
streamlit = ">=1.0.0"
wandb = ">=0.10.31"
matplotlib = ">=3.1.0"
defusedxml = ">=0.7.1,<0.8.0"
opencv-python-headless = ">=4.11.0.86,<5.0.0.0"
anyascii = ">=0.3.2,<0.4.0"
ipykernel = "==6.29.5"
numpy = ">=2.2.5,<3.0.0"
pyyaml = "^6.0.2"
wget = "^3.2"

[tool.poetry.group.dev.dependencies]
pandas = "^2.2.3"
flake8 = "^7.2.0"
pre-commit = "^4.2.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
